fix(api): move composables to separate folder
